Implemented Ryosuke839's quick load patches
Implemented Ryosuke839's quick load patches and more orginized quick
load patches
Fixed game crash on american v1.0 .exe
Added support for SA-MP 0.3x-r2-pre1 and 0.3x-r2-pre2

+++ crashes/crashes/CVector.h
@@ -92,7 +92,7 @@ public:
         vecRotation.fZ = atan2 ( fY, fX );
         CVector vecTemp ( sqrt ( fX * fX + fY * fY ), fZ, 0 );
         vecTemp.Normalize ();
-        vecRotation.fY = atan2 ( vecTemp.fX, vecTemp.fY ) - M_PI / 2;
+        (float)vecRotation.fY = (float)(atan2 ( vecTemp.fX, vecTemp.fY ) - M_PI / 2.0);
         return vecRotation;
     }
